triton-models-inference/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.1.0+cu121
torchvision==0.16.0+cu121
torchaudio==2.1.0+cu121
numpy>=1.21.0,<2.0
opencv-python>=4.5.0
Pillow>=8.0.0
matplotlib>=3.5.0
ultralytics>=8.0.0
